requirements.txt
python-dotenv==0.20.0
requests==2.28.1
python-telegram-bot==13.0.0